apps/backend-blood-analysis/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
tensorflow
keras
scikit-learn
joblib
opencv-python
easyocr
PyMuPDF
numpy
openai
